requirements.txt
imutils==0.4.5
matplotlib==2.1.1
numpy==1.13.3
opencv-python==3.3.0.10
Pillow==4.3.0
pyserial==3.4
scipy==1.0.0
